Refactor `PageHandler`, move initialization functions from `main`
